fix: rename tag payloads folder to reference
