+++ packages/payloads/src/Domain/Payload/PayloadContent.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { ContentReference } from '../Tag/ContentReference'
+import { ContentReference } from '../Reference/ContentReference'
 
 export type PayloadContent = {
   [key: string]: any
