+++ packages/payloads/src/Domain/index.ts
@@ -23,11 +23,11 @@ export * from './Payload/PayloadSource'
 export * from './Payload/PurePayload'
 export * from './Payload/RawPayload'
 export * from './Payload/SurePayload'
-export * from './Tag/AnonymousReference'
-export * from './Tag/ContentReference'
-export * from './Tag/ContenteReferenceType'
-export * from './Tag/Functions'
-export * from './Tag/LegacyAnonymousReference'
-export * from './Tag/LegacyTagToNoteReference'
-export * from './Tag/Reference'
-export * from './Tag/TagToParentTagReference'
+export * from './Reference/AnonymousReference'
+export * from './Reference/ContentReference'
+export * from './Reference/ContenteReferenceType'
+export * from './Reference/Functions'
+export * from './Reference/LegacyAnonymousReference'
+export * from './Reference/LegacyTagToNoteReference'
+export * from './Reference/Reference'
+export * from './Reference/TagToParentTagReference'
